+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 - Tweak: Less cursor type switching
 - Improve styling of Fullscreen-Attachment buttons
 - move the verified icon over the group avatar
+- now it is possible to have multiple accounts with the same email address (you can import a backup next to the active account)
+- change account path again - now the account folder name isn't tied to the email address anymore.
 
 ### Technical